Modified themeprovider and h5 typography
